+++ NEWS.md
@@ -2,13 +2,18 @@
 
 ## Polars R Package (development version)
 
+### Breaking changes
+
+-   Since most of the methods of `Expr` are now available for `Series`, the
+    experimental `<Series>$expr` subnamespace is removed (#831).
+    Use `<Series>$<method>` instead of `<Series>$expr$<method>`.
+
 ### New features
 
 -   New active bindings `$flags` for `DataFrame` to show the flags used internally
     for each column. The output of `$flags` for `Series` was also improved and now
     contains `FAST_EXPLODE` for `Series` of type `list` and `array` (#809).
--   All subnamespaces that can be used on `Expr` are now available for `Series` 
-    (#819, #828).
+-   Most of `Expr` methods are also available for `Series` (#819, #828, #831).
 -   `as_polars_df()` for `data.frame` is more memory-efficient and new arguments
     `schema` and `schema_overrides` are added (#817).
 -   Use `polars_code_completion_activate()` to enable code suggestions and